% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_options.R
\name{editor_options}
\alias{editor_options}
\title{Options for code editor}
\usage{
editor_options(lineNumbers = TRUE, lineWrapping = TRUE, tabSize = 4,
  indentWithTabs = TRUE, smartIndent = TRUE, autofocus = TRUE, ...)
}
\arguments{
\item{lineNumbers}{Logical. Whether to show line numbers to the left of the editor.}

\item{lineWrapping}{Logical. Whether editor should scroll or wrap for long lines.}

\item{tabSize}{The width of a tab character. Defaults to 4.}

\item{indentWithTabs}{Logical. Whether, when indenting,
the first N*\code{tabSize} spaces should be replaced by N tabs.}

\item{smartIndent}{Whether to use the context-sensitive indentation
that the mode provides (or just indent the same as the line before).}

\item{autofocus}{Can be used to make editor focus itself on initialization.}

\item{...}{Additional parameters, see \url{http://codemirror.net/doc/manual.html#config}.}
}
\value{
a list
}
\description{
Options for code editor
}
