% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_infos.R
\name{get_db_infos}
\alias{get_db_infos}
\title{Get tables and fields from a connection}
\usage{
get_db_infos(conn, schema = NULL)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as returned by \code{dbConnect()}.}

\item{schema}{Schema's name, by default \code{NULL}.}
}
\value{
a named \code{list} where names are tables and values tables' fields.
}
\description{
Get tables and fields from a connection
}
\examples{

library(DBI)
con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbWriteTable(conn = con, name = "mtcars", value = mtcars)
get_db_infos(con)

}
