% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery-proxy.R
\name{sq_proxy_clipboard}
\alias{sq_proxy_clipboard}
\title{Copy Editor's value to clipboard}
\usage{
sq_proxy_clipboard(shinyId, selector,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
editor to modify (if invoked from a Shiny module, the namespace will be added
automatically)}

\item{selector}{A CSS selector, e.g. : \code{".btn"} or \code{"#my-button"}.}

\item{session}{the Shiny session object to which the editor belongs; usually the
default value will suffice}
}
\description{
Attach an event to a button to copy to clipboard.
}
\examples{
\dontrun{

if (interactive()) {

  library(shiny)
  library(shinyWidgets)
  library(sqlquery)

  ui <- fluidPage(
    tags$h2("sqlquery example"),
    panel(
      heading = "SQL editor",
      status = "primary",
      sqlqueryOutput(outputId = "mySQL", height = "200px")
    ),
    verbatimTextOutput(outputId = "res"),
    actionButton(inputId = "copy", label = "Copy to clipboard")
  )

  server <- function(input, output, session) {

    output$mySQL <- renderSqlquery({
      sql_query(raw = TRUE)
    })

    # No need to use an observe handler
    sq_proxy_clipboard("mySQL", selector = "#copy")

    output$res <- renderPrint({
      input$mySQL_value
    })

  }

  runApp(
    appDir = shinyApp(ui, server),
    launch.browser = paneViewer()
  )

}

}
}
