% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.R
\name{sql_query}
\alias{sql_query}
\title{Create a SQL Query editor}
\usage{
sql_query(conn = NULL, schema = NULL, autocomplete = NULL, value = NULL,
  options = editor_options(), font_size = "16px", raw = FALSE,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as returned by \code{dbConnect()}.
Default to \code{NULL}.}

\item{schema}{Schema's name, by default \code{NULL}.}

\item{autocomplete}{A named list used for autocompletion.
List's names represent tables, values tables' fields.}

\item{value}{Character. Value to display by default.}

\item{options}{Editor's options, see \code{\link{editor_options}}.}

\item{font_size}{Code editor font size.}

\item{raw}{Logical. If \code{TRUE} only code editor is displayed.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Create a SQL Query editor
}
\examples{
\dontrun{

# Launch default editor
sql_query(value = "SELECT * FROM mtcars")

# Use a connection
con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbWriteTable(conn = con, name = "mtcars", value = mtcars)

sql_query(conn = con)

}
}
