% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_addin.R
\name{sql_query_addin}
\alias{sql_query_addin}
\title{SQL Editor Addin}
\usage{
sql_query_addin(conn = NULL, schema = NULL)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as returned by \code{dbConnect()}.}

\item{schema}{Schema's name, by default \code{NULL}.}
}
\description{
Write SQL queries and execute them.
}
\examples{
\dontrun{

if (interactive()) {

sql_query_addin()

}

}
}
